Modificar el Main Crear una nueva ventana para dar de alta a un Socio Crear un metodo para menuBar Gestion de los metodos para videoclub
